Unit test to test the entire charging framework

+++ logic/src/main/scala/gr/grnet/aquarium/logic/accounting/policies/DefaultRatePolicy.scala
@@ -4,7 +4,7 @@ import gr.grnet.aquarium.logic.accounting.{AccountingEvent, AccountingEntryType,
 
 class DefaultRatePolicy(et: AccountingEntryType.Value) extends Policy(et) {
 
-  def calculateAmount(evt: AccountingEvent) : Double = {
+  def calculateAmount(evt: AccountingEvent) : Float = {
     evt.value() * evt.getRate()
   }
 }
